+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Changelog
  * added `puli self-update` command
  * added `--parsed` option to `puli config`
  * added `--format` option to `puli package`
+ * added `puli package rename` command
 
 * 1.0.0-beta4 (2015-04-13)